Factor out project.util.instance_change_tracker.
